refactor private functions
update reconstruction notebook with new dimension names

refactor pseudo_ens and make public function
